requirements/base.txt
cssmin==0.2.0
dj-database-url==0.3.0
django-braces==1.4.0
django-dotenv==1.2
django-model-utils==2.0.3
django-pipeline==1.3.24
Django==1.6.5
djorm-ext-pool==0.8.1
futures==2.1.6
jsmin==2.0.9
psycopg2==2.5.3
py-bcrypt==0.4
six==1.6.1
South==0.8.4
SQLAlchemy==0.9.4
